create_distribution

+++ pallets/distribution/src/mock.rs
@@ -35,7 +35,8 @@ impl frame_system::Config for Test {
 	type Nonce = u64;
 	type Hash = H256;
 	type Hashing = BlakeTwo256;
-	type AccountId = u64;
+	// Larger `AccountId` type to handle `stash_account` generation.
+	type AccountId = u128;
 	type Lookup = IdentityLookup<Self::AccountId>;
 	type Block = Block;
 	type RuntimeEvent = RuntimeEvent;
